share method
