+++ src/gene/etl/ensembl.py
@@ -14,7 +14,6 @@ from gene.schemas import (
     DataLicenseAttributes,
     NamespacePrefix,
     SourceMeta,
-    StoredSequenceLocation,
     Strand,
 )
 
@@ -88,36 +87,6 @@ class Ensembl(Base):
         gene_params["item_type"] = "identity"
 
         return gene_params
-
-    # TODO reincorporate with NCBI
-    def _build_sequence_location(
-        self, seq_id: str, row: pd.Series, concept_id: str
-    ) -> Optional[StoredSequenceLocation]:
-        """Construct a sequence location for storing in a DB.
-
-        :param seq_id: The sequence ID.
-        :param row: A gene from the source file.
-        :param concept_id: record ID from source
-        :return: A storable SequenceLocation containing relevant params for returning a
-        VRS SequenceLocation, or None if unable to retrieve valid parameters
-        """
-        aliases = self._get_seq_id_aliases(seq_id)
-        if not aliases or row.start is None or row.end is None:
-            return None
-
-        sequence = aliases[0]
-
-        if row.start != "." and row.end != "." and sequence:
-            if 0 <= row.start <= row.end:
-                return StoredSequenceLocation(
-                    start=row.start - 1,
-                    end=row.end,
-                    sequence_id=sequence,
-                )
-            else:
-                _logger.warning(
-                    f"{concept_id} has invalid interval: start={row.start - 1} end={row.end}"
-                )
 
     def _add_attributes(self, row: pd.Series, gene: Dict) -> None:
         """Add concept_id, symbol, and xrefs to a gene record.
